Implemented About Page with responsive UI
