+++ package.json
@@ -10,6 +10,7 @@
     "react-dom": "^18.2.0",
     "react-lottie": "^1.2.4",
     "react-reader": "^2.0.9",
+    "react-router-dom": "^6.30.0",
     "react-scripts": "^5.0.1",
     "web-vitals": "^2.1.4"
   },
