+++ src/pages/About.jsx
@@ -1,0 +1,16 @@
+import React from "react";
+
+const About = () => {
+  return (
+    <div className="container mx-auto p-6">
+      <h1 className="text-4xl font-bold text-center">About Us</h1>
+      <p className="text-center mt-4">We are dedicated to bringing the best services to our customers.</p>
+      <section className="mt-8">
+        <h2 className="text-2xl font-semibold">Our Mission</h2>
+        <p>To revolutionize the industry with cutting-edge technology.</p>
+      </section>
+    </div>
+  );
+};
+
+export default About;
