+++ src/pages/Services.jsx
@@ -1,0 +1,16 @@
+import React from "react";
+
+const Services = () => {
+  return (
+    <div className="container mx-auto p-6">
+      <h1 className="text-4xl font-bold text-center">Our Services</h1>
+      <ul className="mt-6">
+        <li className="text-lg">🔹 Service 1: High-quality solutions</li>
+        <li className="text-lg">🔹 Service 2: Customer-focused approach</li>
+      </ul>
+      <button className="mt-6 bg-blue-600 text-white px-4 py-2 rounded">Get Started</button>
+    </div>
+  );
+};
+
+export default Services;
